requirements.txt
altgraph==0.16.1
awscli==1.16.44
boto3==1.9.34
botocore==1.12.34
colorama==0.3.9
docutils==0.14
future==0.17.1
jmespath==0.9.3
macholib==1.11
pefile==2018.8.8
pyasn1==0.4.4
PyInstaller==3.4
pymssql==2.1.4
python-dateutil==2.7.5
pywin32-ctypes==0.2.0
PyYAML==3.13
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
urllib3==1.24
